Rename to SimulationTimeSeriesMatrix
