+++ frontend/src/modules/SimulationTimeSeriesMatrix/preview.tsx
