+++ frontend/src/modules/SimulationTimeSeriesMatrix/queryHooks.tsx
