+++ frontend/src/modules/SimulationTimeSeriesMatrix/state.ts
