+++ frontend/src/modules/SimulationTimeSeriesMatrix/utils/ColorUtils/colors.ts
