+++ frontend/src/modules/SimulationTimeSeriesMatrix/utils/PlotFigureBuilder/plotFigureBuilderBase.ts
