+++ frontend/src/modules/SimulationTimeSeriesMatrix/utils/PlotFigureBuilder/vectorSubplotBuilder.ts
