+++ frontend/src/modules/SimulationTimeSeriesMatrix/utils/PlotlyTraceUtils/timeSeriesPlotData.ts
