+++ frontend/src/modules/SimulationTimeSeriesMatrix/utils/elementOptionsUtils.ts
