+++ frontend/src/modules/SimulationTimeSeriesMatrix/utils/plotUtils.ts
