+++ frontend/src/modules/SimulationTimeSeriesMatrix/utils/stringUtils.ts
