+++ frontend/src/modules/SimulationTimeSeriesMatrix/view.tsx
