+++ frontend/src/modules/registerAllModules.ts
@@ -10,11 +10,11 @@ import "./Map/registerModule";
 import "./Pvt/registerModule";
 import "./Sensitivity/registerModule";
 import "./SimulationTimeSeries/registerModule";
-import "./SimulationTimeSeriesNew/registerModule";
-import "./WellCompletion/registerModule"
+import "./SimulationTimeSeriesMatrix/registerModule";
 import "./SimulationTimeSeriesSensitivity/registerModule";
 import "./TimeSeriesParameterDistribution/registerModule";
 import "./TopographicMap/registerModule";
+import "./WellCompletion/registerModule";
 
 if (isDevMode()) {
     await import("./MyModule/registerModule");
